+++ framework/ui/components/atoms/vc-col/index.ts
@@ -1,10 +1,1 @@
-import { VNode } from "vue";
-import _Col from "./vc-col.vue";
-
-export const VcCol = _Col as typeof _Col & {
-  new (): {
-    $slots: {
-      default: () => VNode[];
-    };
-  };
-};
+export { default as VcCol } from "./vc-col.vue";