+++ framework/ui/components/atoms/vc-row/index.ts
@@ -1,10 +1,1 @@
-import { VNode } from "vue";
-import _Row from "./vc-row.vue";
-
-export const VcRow = _Row as typeof _Row & {
-  new (): {
-    $slots: {
-      default: () => VNode[];
-    };
-  };
-};
+export { default as VcRow } from "./vc-row.vue";