+++ framework/ui/components/atoms/vc-row/vc-row.vue
@@ -4,10 +4,17 @@
   </div>
 </template>
 
+<script setup lang="ts">
+defineSlots<{
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  default: (props: any) => any;
+}>();
+</script>
+
 <style lang="scss">
 .vc-row {
   .vc-app_mobile & {
-    display: block;
+    display: grid;
   }
 }
 </style>